Back/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
Django==2.2
django-cors-headers==3.5.0
django-cors-middleware==1.5.0
djangorestframework==3.12.2
gunicorn==20.0.4
importlib-metadata==2.0.0
iniconfig==1.1.1
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytz==2020.4
setuptools==47.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
zipp==3.4.0